Refactor QuizQuestionType and QuizQuestion models to change Id and TypeId properties from long to int for improved consistency and data handling.

+++ Chik.Exams/src/Modules/Quiz/Questions/Models/QuizQuestion.cs
@@ -4,7 +4,7 @@ public record QuizQuestion(
     long Id,
     long QuizId,
     string Prompt,
-    long TypeId,
+    int TypeId,
     string Properties,
     int Score,
     int Order,
